Use `uint8_t` instead of `int` for `bits`
The `bits` field can only contain values between 0-7. Replacing `int`
with `uint8_t` to avoid any potential bugs.

+++ c/indexer.h
@@ -66,7 +66,7 @@ struct gzip_index_point
 {
     off_t out;          /* corresponding offset in uncompressed data */
     off_t in;           /* offset in input file of first full byte */
-    int bits;           /* number of bits (1-7) from byte at in - 1, or 0 */
+    uint8_t bits;           /* number of bits (1-7) from byte at in - 1, or 0 */
     unsigned char window[WINSIZE];  /* preceding 32K of uncompressed data */    
 };
